Handle strings correctly.
Note that we did not use string_literals in the setup.py. That is
because Python 2 requires that the setup packages argument provide
a list of bytes, but Python 3 requires a list of strs. If
string_literals are used, you must mark the strings with ``b``
for it to work in Python 2, but that makes it fail in Python 3.